Before refactoring simplebasket

+++ Core/Constants.cs
@@ -2,5 +2,5 @@ namespace Core;
 
 public interface IConstants
 {
-    public static readonly decimal TaxRate = (decimal)0.05;
+    public static readonly decimal TaxRate = 0.05m;
 }
